Delete: フォーマット選択を削除
