+++ package.json
@@ -65,7 +65,6 @@
     "vue": "3.4.26",
     "vuedraggable": "4.1.0",
     "vuex": "4.0.2",
-    "wasm-media-encoders": "0.7.0",
     "zod": "3.22.4"
   },
   "optionalDependencies": {
