+++ src/backend/electron/preload.ts
@@ -54,11 +54,10 @@ const api: Sandbox = {
     return await ipcRendererInvokeProxy.GET_ALT_PORT_INFOS();
   },
 
-  showAudioSaveDialog: ({ title, defaultPath, formats }) => {
+  showAudioSaveDialog: ({ title, defaultPath }) => {
     return ipcRendererInvokeProxy.SHOW_AUDIO_SAVE_DIALOG({
       title,
       defaultPath,
-      formats,
     });
   },
